requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
chardet==3.0.4
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
discord==2.3.2
discord.py==2.3.2
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pdatetime==0.1
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
wcwidth==0.2.12
yarl==1.9.3
